requirements.txt
PySide6>=6.6.0
qt-material>=2.14.0
pillow>=10.3.0
